+++ app/components/marty/diagnostic/report_view/client/report_view.js
@@ -1,0 +1,10 @@
+({
+  initComponent() {
+    this.callParent();
+    this.getView().addElListener("keydown", (e) => {
+      if (e.keyCode === e.ESC) {
+        this.getSelectionModel().deselectAll();
+      }
+    });
+  }
+});