+++ app/models/marty/diagnostic/report.rb
@@ -1,0 +1,8 @@
+module Marty
+  module Diagnostic
+    class Report < Marty::Base
+      self.table_name = :marty_diagnostic_reports
+      has_many :configurations, dependent: :nullify
+    end
+  end
+end