+++ db/migrate/619_create_marty_diagnostic_reports.rb
@@ -1,0 +1,7 @@
+class CreateMartyDiagnosticReports < ActiveRecord::Migration[6.0]
+  def change
+    create_table :marty_diagnostic_reports do |t|
+      t.string :name, null: false
+    end
+  end
+end