+++ lib/marty/diagnostic/fatal.rb
@@ -1,8 +1,13 @@
-module Marty::Diagnostic; class Fatal < Base
-  def self.message(msg, opts = {})
-    node = opts[:node] || Node.my_ip
-    type = opts[:type] || 'RuntimeError'
-    { name.demodulize => { node => { type => error(msg) } } }
+module Marty
+  module Diagnostic
+    class Fatal < Base
+      class << self
+        def message(msg, opts = {})
+          node = opts[:node] || Node.my_ip
+          type = opts[:type] || 'RuntimeError'
+          { name.demodulize => { node => { type => error(msg) } } }
+        end
+      end
+    end
   end
-end
 end